# docs/swagger.yaml
basePath: /
host: localhost:8080
info:
  contact:
    email: soberkoder@swagger.io
    name: API Support
  description: simple api for chat in rooms
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: chat app
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: LogIn to an account
      parameters:
      - description: email of the user
        in: formData
        name: email
        required: true
        type: string
      - description: some password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: LogIn
      tags:
      - Users
  /logout:
    post:
      consumes:
      - application/json
      description: Logout from an account
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Logout
      tags:
      - Users
  /register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: username
        in: formData
        name: username
        required: true
        type: string
      - description: email of the user
        in: formData
        name: email
        required: true
        type: string
      - description: some password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Register
      tags:
      - Users
swagger: "2.0"